refactor: split circular progress bar
